// be/src/requests/enums/request-type.enum.ts
export enum RequestType {
    PAID_LEAVE = 'PAID_LEAVE',
    UNPAID_LEAVE = 'UNPAID_LEAVE',
    LEAVE = 'LEAVE',
    OVERTIME = 'OVERTIME',
    BUSINESS_TRIP = 'BUSINESS_TRIP',
    RESIGNATION = 'RESIGNATION',
    ATTENDANCE_ADJUSTMENT = 'ATTENDANCE_ADJUSTMENT',
}

export enum RequestStatus {
    PENDING = 'PENDING',
    APPROVED = 'APPROVED',
    REJECTED = 'REJECTED',
}